[move/#34] JobStepStatus.kt Type 폴더로 이동
