+++ app/src/main/java/com/haphap/app/presentation/jobdetail/component/JobPostStep.kt
@@ -7,6 +7,7 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -16,6 +17,8 @@ import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.haphap.app.core.designsystem.theme.HapHapTheme
+import com.haphap.app.presentation.jobdetail.type.JobStepStatus
+import kotlinx.collections.immutable.ImmutableList
 
 
 @Composable
@@ -83,6 +86,26 @@ fun JobPostStep(
             style = stateTextStyle,
             color = stateTextColor,
         )
+    }
+}
+
+@Composable
+fun JobPostStepRow(
+    steps: ImmutableList<JobStep>,
+    modifier: Modifier = Modifier,
+) {
+    LazyRow(
+        modifier = modifier,
+        horizontalArrangement = Arrangement.spacedBy(12.dp),
+    ) {
+        items(steps) { step ->
+            JobPostStep(
+                number = step.number,
+                stageName = step.stageName,
+                stateText = step.stateText,
+                status = step.status,
+            )
+        }
     }
 }
