+++ app/src/main/java/com/haphap/app/presentation/jobdetail/type/JobStepStatus.kt
@@ -1,4 +1,5 @@
-package com.haphap.app.presentation.jobdetail.component
+package com.haphap.app.presentation.jobdetail.type
+
 enum class JobStepStatus {
     COMPLETED,
     IN_PROGRESS,
